Eliminate circular dependency
This was preventing the Exchange wrapper from ever importing its
validator!
