+++ python-packages/json_schemas/setup.py
@@ -162,7 +162,6 @@ setup(
     extras_require={
         "dev": [
             "0x-contract-addresses",
-            "0x-contract-wrappers",
             "bandit",
             "black",
             "coverage",
