+++ python-packages/sra_client/src/zero_ex/sra_client/__init__.py
@@ -105,7 +105,8 @@ Post Order
 
 Post an order for our Maker to trade ZRX for WETH:
 
->>> from zero_ex.contract_wrappers.exchange.types import Order, order_to_jsdict
+>>> from zero_ex.contract_wrappers.exchange.types import Order
+>>> from zero_ex.contract_wrappers.order_conversions import order_to_jsdict
 >>> from zero_ex.order_utils import (
 ...     asset_data_utils,
 ...     sign_hash)
